Suckers & annoying bugs
Addition of a rudimentary Sucker node. As its name cleverly implies, it sucks nearby objects. A solution to combine RigidBodies may have been found. However, most of the time was spent on the meshCurver GDNative module, that still doesn't work. Reason of crash may be a bug in godot-cpp ? It was submitted as an issue, hopefully something can be worked out.